apps/berg10-agent/pyproject.toml
[project]
name = "berg10-agent"
version = "0.1.0"
description = "Headless LLM agent with WebSocket transport using PocketFlow and LiteLLM"
requires-python = ">=3.10"
dependencies = [
    "pocketflow>=0.0.3",
    "litellm>=1.83.11",
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.45.0",
    "python-dotenv>=1.2.2",
    "pydantic>=2.13.3",
    "click>=8.3.2",
    "httpx>=0.28.1",
    "jsonschema>=4.26.0",
    "aiofiles>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "black>=26.3.1",
    "ruff>=0.15.11",
]
ui = [
    "streamlit>=1.56.0",
]

[project.scripts]
berg10-agent = "berg10_agent.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/berg10_agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
